==> code/2_EDA.ipynb <==
# import libraries

import pandas as pd 
import numpy as np 
import matplotlib.cm as cm
import matplotlib.colors as colors
import matplotlib.pyplot as plt
import seaborn as sns 

# Read in dataframe

venues_df = pd.read_csv('../data/venues_df.csv')
venues_df.head()

# Read in dataframe 

data_df = pd.read_csv('../data/data_df.csv')
data_df.head()

data_df.describe()

data_df.shape[0]

price_groups = []

for price in data_df['avg_price_per_sqft']:
    if price >= 0 and price < 400:
        group = 0
        price_groups.append(group)
    else:
        group = 1
        price_groups.append(group)

# Add price groups to dataframe
data_df['price_group'] = price_groups

# Save dataframe with price groups for Tableau 
data_df.to_excel('../data/price_groups.xlsx', index=False)

data_df.head()

# View value counts distribution

data_df['price_group'].value_counts()

# Confirm number of remaining neighborhoods

len(venues_df['neighborhood'].unique())

# Preview
venues_df.head()

# Count of unique venues

len(venues_df['venue_type'].unique())

# drop unncessary columns

venues_df.drop(columns=['latitude', 'longitude', 'venue_name', 'venue_latitude', 'venue_longitude'], inplace=True)
venues_df.head()

# Count of venues per neighborhood

venues_df.groupby('neighborhood').count().head()

# Dummify venue types

df_onehot = pd.get_dummies(venues_df[['venue_type']], prefix='', prefix_sep='')

# Add neighborhoods to df
df_onehot['neighborhood'] = venues_df['neighborhood'] 

# move neighborhood column to the first column
fixed_columns = [df_onehot.columns[-1]] + list(df_onehot.columns[:-1])
df_onehot = df_onehot[fixed_columns]

# Preview

df_onehot.head()

# Check new shape

df_onehot.shape

# Group venue type count by neighborhood

df_grouped = df_onehot.groupby('neighborhood').sum().reset_index() 

# Preview
df_grouped.head()

# View size

df_grouped.shape

# Combine dataframes

df_comb = pd.merge(df_grouped, data_df, on='neighborhood')
df_comb.head()

# Drop unnecessary columns

df_comb.drop(columns=['location', 'latitude', 'longitude', 'url_format', 'url', 'median_prices', 'avg_price_per_sqft'], inplace=True)
df_comb.head()

# Save dataframe for modeling 

df_comb.to_csv('../data/df_comb.csv', index=False)
