==> code/3_Modeling.ipynb <==
# import libraries

import pandas as pd 
import numpy as np 
import matplotlib.cm as cm
import matplotlib.colors as colors
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler, normalize, scale
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score, cross_val_predict
from sklearn.linear_model import LinearRegression, LassoCV, RidgeCV
from sklearn.decomposition import PCA
from sklearn import svm
from sklearn.metrics import mean_squared_error, r2_score, accuracy_score
from sklearn.tree import DecisionTreeClassifier
from sklearn.linear_model import LogisticRegression

# Read in dataframe

df_comb = pd.read_csv('../data/df_comb.csv')
df_comb.head()

# Modeling

#### Prepocessing

# Define X and y

X = df_comb.drop(columns=['neighborhood', 'price_group'])
y = df_comb['price_group']

# Train/test split 

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)

# # Scale data

sc = StandardScaler()
X_train = sc.fit_transform(X_train)
X_test = sc.transform(X_test)

#### PCA Modeling

# Instantiate and fit model
pca = PCA(svd_solver='auto', random_state=0)
# X_pca = pca.fit(X_train) 
X_pca = pca.fit_transform(scale(X))

# Code adapted from Medium article on PCA 

n_component_list = range(1, 51)
R2_list = []
MSE_list = []

# Linear Regression
for i in n_component_list:
    lr = LinearRegression()
    X_train, X_test, y_train, y_test = train_test_split(X_pca[:,:i], y, test_size=0.2, random_state=0)
    model = lr.fit(X_train, y_train)
    # check the result
    y_pred = lr.predict(X_test)
    r2 = r2_score(y_test, y_pred) # r2 score
    mse = mean_squared_error(y_test, y_pred) # mse
    R2_list.append(r2)
    MSE_list.append(mse)
    
scores_df = pd.DataFrame.from_dict(dict([('NComponents', n_component_list),
                                        ('R2', R2_list),
                                        ('MSE', MSE_list)]))
scores_df.set_index('NComponents', inplace=True)

# View best scores

r2_max = scores_df['R2'].idxmax()
print("Best n:", r2_max, ", R2 score:", scores_df['R2'][r2_max])

mse_min = scores_df['MSE'].idxmin()
print("Best n:", mse_min, ", MSE:", scores_df['MSE'][mse_min])

# Use the best n_components parameter

# Instantiate and fit
lr = LinearRegression()
X_train, X_test, y_train, y_test = train_test_split(X_pca[:,:r2_max], y, test_size=0.2, random_state=0)
model = lr.fit(X_train, y_train)

# View scores

y_pred = lr.predict(X_test)
r2 = r2_score(y_test, y_pred) # r2 score
mse = mean_squared_error(y_test, y_pred) # mse
print("R2 score:", r2)
print("MSE:", mse)

# Project the coefs back to the original number of features

eigenvectors = pca.components_
pcr_coefs = eigenvectors[:r2_max, :].T @ lr.coef_

# View with venue types have most and least effect

print('\nMax positive coefs:', pcr_coefs[np.argsort(-pcr_coefs)[:5]])
print('\nVenue types with most positive effect:', X.columns[np.argsort(-pcr_coefs)[:5]].values)
print('\nMax negative coefs:', pcr_coefs[np.argsort(pcr_coefs)[:5]])
print('\nVenue types with most negative effect:', X.columns[np.argsort(pcr_coefs)[:5]].values)
coef_abs = abs(pcr_coefs)
print('\nMin coefs:', pcr_coefs[np.argsort(coef_abs)[:5]])
print('\nVenue types with least effect:', X.columns[np.argsort(coef_abs)[:5]].values)

#### SVM Model

# Instantiate and fit model

svm_model = svm.SVC()
svm_model.fit(X_train, y_train) 

# Generate predictions

y_pred = svm_model.predict(X_test)
y_pred

# Get accuracy scores

print("SVM Training Accuracy Score: %.3f" % svm_model.score(X_train, y_train))
print("SVM Testing Accuracy Score: %.3f" % svm_model.score(X_test, y_test)) 

#### Decision Tree

# Instantiate and fit model

dt_model = DecisionTreeClassifier(criterion="entropy", max_depth = 4)
dt_model.fit(X_train,y_train)
# dt_model

# Generate predictions

y_pred = dt_model.predict(X_test)
y_pred

print("Decision Tree Training Accuracy Score: %.3f" % dt_model.score(X_train, y_train))
print("Decision Tree Testing Accuracy Score: %.3f" % dt_model.score(X_test, y_test))

#### Logistic Regression

# Instantiate and fit model

logreg_model = LogisticRegression(C=0.01).fit(X_train,y_train)
logreg_model

# Generate predictions

y_pred = logreg_model.predict(X_test)
y_pred

print("Logreg Training Accuracy Score: %.3f" % logreg_model.score(X_train, y_train))
print("Logreg Tree Testing Accuracy Score: %.3f" % logreg_model.score(X_test, y_test))
